refactor: move appendSwitch / appendArgument to command_line module
