+++ lib/browser/api/app.js
@@ -1,6 +1,7 @@
 'use strict'
 
 const bindings = process.atomBinding('app')
+const commandLine = process.atomBinding('command_line')
 const path = require('path')
 const { app, App } = bindings
 
@@ -17,6 +18,12 @@ let dockMenu = null
 Object.setPrototypeOf(App.prototype, EventEmitter.prototype)
 EventEmitter.call(app)
 
+const castArgs = function (args) {
+  return args.map((arg) => {
+    return typeof arg !== 'string' ? `${arg}` : arg
+  })
+}
+
 Object.assign(app, {
   setApplicationMenu (menu) {
     return Menu.setApplicationMenu(menu)
@@ -25,19 +32,10 @@ Object.assign(app, {
     return Menu.getApplicationMenu()
   },
   commandLine: {
-    ...process.atomBinding('command_line'),
-    appendSwitch (...args) {
-      const castedArgs = args.map((arg) => {
-        return typeof arg !== 'string' ? `${arg}` : arg
-      })
-      return bindings.appendSwitch(...castedArgs)
-    },
-    appendArgument (...args) {
-      const castedArgs = args.map((arg) => {
-        return typeof arg !== 'string' ? `${arg}` : arg
-      })
-      return bindings.appendArgument(...castedArgs)
-    }
+    hasSwitch: (...args) => commandLine.hasSwitch(...castArgs(args)),
+    getSwitchValue: (...args) => commandLine.getSwitchValue(...castArgs(args)),
+    appendSwitch: (...args) => commandLine.appendSwitch(...castArgs(args)),
+    appendArgument: (...args) => commandLine.appendArgument(...castArgs(args))
   }
 })
